requirements.txt
google-cloud-vision==3.7.2
openai==1.23.6
python-dotenv==1.0.1
Flask==3.0.3
opencv-python==4.9.0.80
Pillow==10.3.0
moviepy==1.0.3